Improved CLI loop & added SaleRow class

+++ fruit.py
@@ -1,16 +1,15 @@
-
 from dataclasses import dataclass
+
 
 @dataclass
 class Fruit:
-
     """
     Classe générique Fruit pouvant retourner la valeur totale d'un fruit restant en stock.
     """
 
     name: str
     stock: int
-    price: int
+    price: float
 
     def stock_value(self) -> float:
         return self.stock * self.price
